chore: restructuring and renaming directories
