+++ packages/amplify-cli/src/commands/gen2-migration/codegen-generate/src/adapters/storage/gen1_storage_codegen_adapter.ts
