+++ packages/amplify-cli/src/commands/gen2-migration/codegen-generate/src/adapters/storage/index.ts
