+++ packages/amplify-cli/src/commands/gen2-migration/codegen-generate/src/backend/synthesizer.ts
@@ -14,7 +14,7 @@ import type { BucketAccelerateStatus, BucketVersioningStatus } from '@aws-sdk/cl
 import { AccessPatterns, ServerSideEncryptionConfiguration } from '../generators/storage/index.js';
 import { ExplicitAuthFlowsType, OAuthFlowType, UserPoolClientType } from '@aws-sdk/client-cognito-identity-provider';
 import assert from 'assert';
-import { newLineIdentifier } from '../ts_factory_utils';
+import { newLineIdentifier } from '../ts_factory_utils.js';
 
 const factory = ts.factory;
 export interface BackendRenderParameters {
