+++ packages/amplify-cli/src/commands/gen2-migration/codegen-generate/src/cli/migration-command.ts
