+++ packages/amplify-cli/src/commands/gen2-migration/codegen-generate/src/core/project-analyzer.ts
