+++ packages/amplify-cli/src/commands/gen2-migration/codegen-generate/src/generators/auth/index.test.ts
