+++ packages/amplify-cli/src/commands/gen2-migration/codegen-generate/src/generators/functions/lambda.ts
