+++ packages/amplify-cli/src/commands/gen2-migration/codegen-generate/src/migration_head/app_storage_definition_fetcher.ts
@@ -1,8 +1,8 @@
 import fs from 'node:fs/promises';
 import path from 'node:path';
-import { getStorageDefinition } from '../../adapters/storage/index';
+import { getStorageDefinition } from '../adapters/storage/index';
 import { BackendDownloader } from './backend_downloader.js';
-import { StorageRenderParameters, StorageTriggerEvent, Lambda, ServerSideEncryptionConfiguration } from '../../core/migration-pipeline';
+import { StorageRenderParameters, StorageTriggerEvent, Lambda, ServerSideEncryptionConfiguration } from '../core/migration-pipeline';
 import {
   GetBucketNotificationConfigurationCommand,
   S3Client,
