+++ packages/amplify-cli/src/commands/gen2-migration/codegen-generate/src/renderers/package_json.ts
