+++ packages/amplify-cli/src/commands/gen2-migration/codegen-generate/src/resource/resource.ts
